Initial [public] commit.
Originally I tracked this code in a private repository. Originally the
Web site also had a very little bit of PHP to transform XSL on the fly
and dynamically generate the 404 output.

It's been a couple of years since the PHP was used. I've been going
static-only for a while to force myself to learn Perl Web frameworks. :)
I have begun to make some progress with Catalyst and badly want to learn
it so hopefully in the near future I will Catalyst this up. In any case,
I'm leaving out the PHP entirely since I never intend to use it again
and it was nothing speical.

There was also some Apache configuration that glued things together
somewhat, and some shell scripts to update and manage the Web server,
and statically transform the XSL (after I stopped using PHP to do it).
The scripts weren't very good and I've basically moved on from using
Apache now. Currently I'm using nginx, but going forward I suspect that
I'll be using Starman instead.

In short, the site is kind of a mess right now. I'm going to need to
write some new scripts for updating and managing the site, and possibly
some Perl to transform the XSL (either statically or at run-time). We
shall see... For the most part there shouldn't be anything revolutionary
here. Most of what remains was publically visble anyway...